Batch compiler Main now uses a special 'aspectj' locale
The fictitious ISO country code 'aspectj' identifies a resource bundle
'messages_aspectj.properties' in the AspectJ code base. That file is
meant to define a delta of AspectJ-specific properties not found in the
original 'messages.properties' or overriding them, such as compiler name
and version. This way we do not have to merge the two files all the
time, keeping their contents in sync. Java will automatically handle the
country-specific resource bundle with precedence over the unspecific one
found in the Eclipse code base.

The code was updated in order to custom locale instead of the system's
default locale. This way 'static String bundleName' can become final
again like in the Eclipse code base. AspectJ no longer needs to
overwrite it in its BuildArgParser during static initialisation.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>