Add Maven build to AspectJ JDT Core
I kept the whole Eclipse Maven POM on the top of the file, but commented
it out. This should make merges from upstream easier because this way
Git can find the right hunks to merge in. the original is part of a
multi-module build, though, and we are not building the full JDT Core
project, only one module.

Please also note my extensive comments and the long list of shading
exclusions in order to get the relocation exceptions right for both
binary and source uber JARs. This should take care of preserving the
config keys in AssistOptions, CompilerOptions (like before in the Ant
build) plus many more references to config keys which I assume should
also remain unchanged throughout the core and APT modules.

Unfortunately, in order to get the replacements in the source uber JAR
in order, too, I had too fix a very old problem in Maven Shade (source
relocations ignore excludes) and am currently building with a snapshot
of that locally until hopefully my PR in Maven Shade will have been
accepted and a new version released. See also:
https://issues.apache.org/jira/browse/MSHADE-252
https://github.com/apache/maven-shade-plugin/pull/88

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>