Use self-built version of Maven Shade with fix for MSHADE-252
I deployed the fix to GitHub Packages, so we do not have to use a
locally built snapshot version anymore in order to get correctly
relocated and filtered source JARs.

In order to achieve that, I had to add my GitHub Packages repository as
a 'pluginRepository' in the POM temporarily until there is a new
upstream release. Until then, GitHub CI builds should work because they
now know where to find the bugfix version.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>