Flatten shaded JAR plugin, removing shaded dependencies
Using Flatten Maven Plugin goal 'flatten', we can replace the original
POM by a dependency-reduced, streamlined, simple version. The nice thing
about it is that, unlike Maven Shade, Maven Flatten not just creates
such a POM but also attaches it to the build, replacing the original.
Shade just writes it to a directory, doing nothing else with it.

This way both Maven Install and Maven Deploy see the streamlined POM and
we can simply install and deploy without any transitive dependency
headaches.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>