Fix whitespace issues in compiler usage texts
This is a commit cherry-picked from the one I made for this PR:
https://github.com/eclipse/eclipse.jdt.core/pull/62

First and foremost, 'misc.usage.warn' contained tabs instead of spaces,
making assuming a tab width of 4 inside the properties file. But on the
console usually tabs are displayed as 8 spaces. As this is
user-configurable, though, it makes sense to expand the tabs. BTW, not
even when setting the console tab width to 4, output looks correct,
because in the file each continuation line begins with a '\' character,
making the first tab expand to position 4 in the file, but position 3 on
the output line because the backslash is not printed. This makes using
tabs unreliable at best, when trying to predict what the usage text
might look like on the console. Hence, all tabs were expanded to spaces.

In a few other places where indentation for the column layout was also
broken using normal using normal space, I made a few more cosmetic
fixes.

TODO: Add a test verifying that message.properties does not contain any
tab characters or at least leading comments for 'misc.usage' and
'misc.usage.warn' that tabs are to be avoided for layout purposes.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>