Remove remaining tabs from comments in compiler message resource file
This is a preparation for a future automatic verification step during
the build process. If the latter checks for the existence of tab
characters in the file, there should not be any. Therefore, I removed
the remaining ones in the header comment section, replacing them by
spaces. This commit does not change any message texts at all.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>