Exclude more stuff from build artifact (Apache Felix, Eclipse JFace)
Andy noticed that there was stuff inside aspectjtools.jar which in 1.9.6
was not there. I thought it was just new content from JDT dependencies,
but actually Apache Felix and Eclipse JFace existed in the dependencies
before, but were explicitly excluded in the old Ant build. I missed to
port that one line to Maven - good catch, Andy!

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>