Revert Java 8 compatibility code to upstream Java 11
Previously, we managed to keep AJ JDT Core compiling on Java 8, but for
a while this has not been possible anymore because of Java 11 byte code
in other dependencies. I.e., we are building with target 11 anyway.
Therefore, we can remove the previously inserted compatibility code in
favour of better mergeability for upstream changes.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>
Change-Id: I2fe9cf2c09445207f46e8e21907361eea41361ac