Eclipse JDT Core now requires JDK 17
At least for now, we can still use source level 17, but target 11. I.e.,
AJ JDT Core needs to be built on JDK 17+, but for now, AJC should still
build and run on JDK 11.

Some dependencies and POM settings had to be updated in order to fix
source shading and javadoc generation for relocated sources.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>